requirements-test.txt
pytest==7.4.3
pytest-twisted==1.14.0
python-ldap==3.4.4
coverage==7.3.2
